fix: correct the restapi endpoint for task engine submission
Whoops, I had the right functionality under the completely wrong
endpoint!
