+++ src/dioptra/restapi/routes.py
@@ -35,7 +35,6 @@ def register_routes(api: Api, app: Flask) -> None:
     from .experiment import register_routes as attach_experiment
     from .job import register_routes as attach_job
     from .queue import register_routes as attach_job_queue
-    from .task_engine import register_routes as attach_task_engine
     from .task_plugin import register_routes as attach_task_plugin
 
     # Add routes
@@ -43,4 +42,3 @@ def register_routes(api: Api, app: Flask) -> None:
     attach_job(api, app)
     attach_job_queue(api, app)
     attach_task_plugin(api, app)
-    attach_task_engine(api, app)
